refactor pay stories
